Update to ES6 sintax in functions
